chore: move file exporter to settings
