+++ frontend/app_flowy/lib/plugins/document/presentation/more/more_button.dart
@@ -1,8 +1,6 @@
 import 'package:app_flowy/plugins/document/presentation/more/cubit/document_appearance_cubit.dart';
-import 'package:app_flowy/plugins/document/presentation/more/file_exporter_widget.dart';
 import 'package:app_flowy/plugins/document/presentation/more/font_size_switcher.dart';
 import 'package:flowy_infra/image.dart';
-import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -25,24 +23,7 @@ class DocumentMoreButton extends StatelessWidget {
               child: const FontSizeSwitcher(),
             ),
           ),
-          const PopupMenuItem(
-            value: 2,
-            enabled: true,
-            child: FlowyText.medium('Export DataBase'),
-          ),
         ];
-      },
-      onSelected: (value) {
-        if (value == 2) {
-          showDialog(
-            context: context,
-            builder: (context) {
-              return const FlowyDialog(
-                child: FileExporterWidget(),
-              );
-            },
-          );
-        }
       },
       child: svgWidget(
         'editor/details',
