+++ frontend/app_flowy/lib/workspace/presentation/settings/widgets/settings_file_system_view.dart
@@ -1,5 +1,3 @@
-import 'package:app_flowy/plugins/document/presentation/more/file_exporter_widget.dart';
-import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -7,7 +5,9 @@ import 'package:app_flowy/generated/locale_keys.g.dart';
 import 'package:app_flowy/startup/tasks/prelude.dart';
 import 'package:app_flowy/workspace/application/settings/settings_location_cubit.dart';
 import 'package:app_flowy/workspace/presentation/home/toast.dart';
+import 'package:app_flowy/workspace/presentation/settings/widgets/settings_file_exporter_widget.dart';
 import 'package:easy_localization/easy_localization.dart';
+import 'package:file_picker/file_picker.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flowy_infra_ui/style_widget/icon_button.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
